# openapi.yaml
openapi: 3.1.0
info:
  title: Book local API
  version: 0.1.0
paths:
  /health:
    get:
      summary: Health
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /workflows/{name}/runs:
    post:
      summary: Start Run
      operationId: start_run_workflows__name__runs_post
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          title: Name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /runs/{run_id}:
    get:
      summary: Get Run
      operationId: get_run_runs__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /runs/{run_id}/logs:
    get:
      summary: Get Logs
      operationId: get_logs_runs__run_id__logs_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: cursor
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Cursor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunLogsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books:
    get:
      summary: Get Books
      operationId: get_books_books_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooksResponse'
  /books/{book_id}:
    get:
      summary: Get Book
      operationId: get_book_books__book_id__get
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: string
          title: Book Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/{book_id}/runs:
    get:
      summary: Get Book Runs
      operationId: get_book_runs_books__book_id__runs_get
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: string
          title: Book Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/{book_id}/files:
    get:
      summary: Get Files
      operationId: get_files_books__book_id__files_get
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: string
          title: Book Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/{book_id}/files/{path}:
    get:
      summary: Get File
      operationId: get_file_books__book_id__files__path__get
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: string
          title: Book Id
      - name: path
        in: path
        required: true
        schema:
          type: string
          title: Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/{book_id}/voice-assignments:
    patch:
      summary: Patch Voice Assignments
      operationId: patch_voice_assignments_books__book_id__voice_assignments_patch
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: string
          title: Book Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Assignments
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Patch Voice Assignments Books  Book Id  Voice Assignments
                  Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    BookDetail:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        chapters:
          items:
            $ref: '#/components/schemas/ChapterSummary'
          type: array
          title: Chapters
        characters:
          items:
            $ref: '#/components/schemas/CharacterInfo'
          type: array
          title: Characters
        voice_assignments:
          additionalProperties:
            type: string
          type: object
          title: Voice Assignments
      type: object
      required:
      - id
      - title
      - chapters
      - characters
      - voice_assignments
      title: BookDetail
      description: A book's metadata, chapters, characters, and recorded voice assignments.
    BookSummary:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        release_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Release Date
        chapters:
          type: integer
          title: Chapters
        characters:
          type: integer
          title: Characters
        cast:
          type: integer
          title: Cast
      type: object
      required:
      - id
      - title
      - chapters
      - characters
      - cast
      title: BookSummary
      description: A book's headline metadata and casting counts for the list view.
    BooksResponse:
      properties:
        books:
          items:
            $ref: '#/components/schemas/BookSummary'
          type: array
          title: Books
      type: object
      required:
      - books
      title: BooksResponse
      description: The known books with their metadata.
    ChapterSummary:
      properties:
        number:
          type: integer
          title: Number
        title:
          type: string
          title: Title
          default: ''
        beats:
          type: integer
          title: Beats
          default: 0
      type: object
      required:
      - number
      title: ChapterSummary
      description: One chapter's number, title, and how many beats have been extracted.
    CharacterInfo:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        gender:
          anyOf:
          - type: string
          - type: 'null'
          title: Gender
        age:
          anyOf:
          - type: string
          - type: 'null'
          title: Age
        accent:
          anyOf:
          - type: string
          - type: 'null'
          title: Accent
      type: object
      required:
      - id
      - name
      title: CharacterInfo
      description: A character and the voice traits the AI assigned.
    FilesResponse:
      properties:
        files:
          items:
            type: string
          type: array
          title: Files
      type: object
      required:
      - files
      title: FilesResponse
      description: Relative paths of a book's artifact files.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HealthResponse:
      properties:
        status:
          type: string
          title: Status
      type: object
      required:
      - status
      title: HealthResponse
      description: Liveness signal for the platform.
    RunLogsResponse:
      properties:
        lines:
          items:
            type: string
          type: array
          title: Lines
        cursor:
          type: integer
          title: Cursor
        done:
          type: boolean
          title: Done
      type: object
      required:
      - lines
      - cursor
      - done
      title: RunLogsResponse
      description: A page of run log lines and the cursor to fetch the next page.
    RunRequest:
      properties:
        url:
          type: string
          title: Url
        start_chapter:
          type: integer
          title: Start Chapter
          default: 1
        end_chapter:
          anyOf:
          - type: integer
          - type: 'null'
          title: End Chapter
        refresh:
          type: boolean
          title: Refresh
          default: false
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
        book_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Book Id
      type: object
      required:
      - url
      title: RunRequest
      description: Body for starting a workflow run.
    RunStatusResponse:
      properties:
        run_id:
          type: string
          title: Run Id
        workflow:
          type: string
          title: Workflow
        state:
          type: string
          title: State
        returncode:
          anyOf:
          - type: integer
          - type: 'null'
          title: Returncode
        started_at:
          type: string
          title: Started At
          default: ''
        ended_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Ended At
      type: object
      required:
      - run_id
      - workflow
      - state
      title: RunStatusResponse
      description: The state of a workflow run.
    RunSummary:
      properties:
        run_id:
          type: string
          title: Run Id
        workflow:
          type: string
          title: Workflow
        start_chapter:
          anyOf:
          - type: integer
          - type: 'null'
          title: Start Chapter
        end_chapter:
          anyOf:
          - type: integer
          - type: 'null'
          title: End Chapter
        state:
          type: string
          title: State
        started_at:
          type: string
          title: Started At
          default: ''
        ended_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Ended At
      type: object
      required:
      - run_id
      - workflow
      - state
      title: RunSummary
      description: One run's state for a book, so the chapters table can show live
        progress.
    RunsResponse:
      properties:
        runs:
          items:
            $ref: '#/components/schemas/RunSummary'
          type: array
          title: Runs
      type: object
      required:
      - runs
      title: RunsResponse
      description: The runs recorded for a book, oldest first.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError